rename control
