+++ index.ts
@@ -333,7 +333,7 @@ import { Observable } from './src/Observable';
 import { Point } from './src/point.class';
 import { Intersection } from './src/intersection.class';
 import { Color } from './src/color/color.class';
-import { Control } from './src/controls/control.class';
+import { Control } from './src/controls/Control';
 import { Gradient } from './src/gradient/Gradient';
 import { Pattern } from './src/pattern.class';
 import { Shadow } from './src/shadow.class';
