+++ src/controls/polyControl.ts
@@ -1,5 +1,5 @@
 import { Point } from '../point.class';
-import { Control } from './control.class';
+import { Control } from './Control';
 import { TMat2D } from '../typedefs';
 import { iMatrix } from '../constants';
 import type { Polyline } from '../shapes/Polyline';
